adding new interface, redux on unsigned elements
