+++ lib/src/main/kotlin/io/johnnystarr/jvm65/core/Arithmetic.kt
@@ -1,0 +1,6 @@
+package io.johnnystarr.jvm65.core
+
+interface Arithmetic {
+    operator fun plus(increment: Int): UnsignedRegister
+    operator fun minus(decrement: Int): UnsignedRegister
+}
